pyproject.toml
[tool.poetry]
name = "capstone-project"
version = "0.1.0"
description = ""
authors = ["Justin Farnan <77633929+justinm329@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pandas = "2.2.2"
pyarrow = "16.0.0"
numpy = "1.26"
scikit-learn = "1.4.2"
requests = "^2.32.3"
pytz = "^2024.1"
python-dotenv = "^1.0.1"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
